Updating config.yml

+++ resources/mana.sql
@@ -4,15 +4,15 @@
 -- #  { setup
 CREATE TABLE IF NOT EXISTS playerdata(
     playerid TEXT,
-    maxmana INTEGER,
-    manaregen INTEGER
+    maxmana FLOAT,
+    manaregen FLOAT
 )
 -- #  }
 
 -- #  { newdata
 -- #  :playerid string
--- #  :maxmana int
--- #  :manaregen int
+-- #  :maxmana float
+-- #  :manaregen float
 INSERT INTO playerdata(playerid, maxmana, manaregen) VALUES (:playerid, :maxmana, :manaregen)
 -- #  }
 
@@ -23,13 +23,13 @@ SELECT * FROM playerdata WHERE LOWER(playerid) = LOWER(:playerid)
 
 -- #  { setmaxmana
 -- #  :playerid string
--- #  :maxmana int
+-- #  :maxmana float
 UPDATE playerdata SET maxmana = :maxmana WHERE LOWER(playerid) = LOWER(:playerid)
 -- #  }
 
 -- #  { setmanaregen
 -- #  :playerid string
--- #  :manaregen int
+-- #  :manaregen float
 UPDATE playerdata SET manaregen = :manaregen WHERE LOWER(playerid) = LOWER(:playerid)
 -- #  }
 -- #}
